// types/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  timestamptz: any;
  uuid: any;
};
// ...
export type Groups = {
  __typename?: 'groups';
  id: Scalars['uuid'];
  name: Scalars['String'];
  /** fetch data from the table: "users" */
  users: Array<Users>;
  /** An aggregate relationship */
  users_aggregate: Users_Aggregate;
};
// ...
export type Profiles = {
  __typename?: 'profiles';
  id: Scalars['uuid'];
  nickname: Scalars['String'];
  /** An array relationship */
  profiles_users: Array<Profiles_Users>;
  /** An aggregate relationship */
  profiles_users_aggregate: Profiles_Users_Aggregate;
  /** An object relationship */
  user?: Maybe<Users>;
};
// ...
export type Profiles_Users = {
  __typename?: 'profiles_users';
  id: Scalars['uuid'];
  /** An object relationship */
  profile: Profiles;
  profile_id: Scalars['uuid'];
  /** An object relationship */
  user: Users;
  user_id: Scalars['uuid'];
};

/** aggregated selection of "profiles_users" */
export type Profiles_Users_Aggregate = {
  __typename?: 'profiles_users_aggregate';
  aggregate?: Maybe<Profiles_Users_Aggregate_Fields>;
  nodes: Array<Profiles_Users>;
};
// ...
export type Users = {
  __typename?: 'users';
  age?: Maybe<Scalars['String']>;
  created_at: Scalars['timestamptz'];
  /** An object relationship */
  group?: Maybe<Groups>;
  group_id?: Maybe<Scalars['uuid']>;
  id: Scalars['uuid'];
  image?: Maybe<Scalars['String']>;
  job?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  /** An object relationship */
  profile?: Maybe<Profiles>;
  profile_id?: Maybe<Scalars['uuid']>;
  /** An array relationship */
  profiles_users: Array<Profiles_Users>;
  /** An aggregate relationship */
  profiles_users_aggregate: Profiles_Users_Aggregate;
};
// ...
export type Users_Aggregate = {
  __typename?: 'users_aggregate';
  aggregate?: Maybe<Users_Aggregate_Fields>;
  nodes: Array<Users>;
};

/** aggregate fields of "users" */
export type Users_Aggregate_Fields = {
  __typename?: 'users_aggregate_fields';
  count: Scalars['Int'];
  max?: Maybe<Users_Max_Fields>;
  min?: Maybe<Users_Min_Fields>;
};
// ...
export type GetGroupsQuery = { __typename?: 'query_root', groups: Array<{ __typename?: 'groups', id: any, name: string, users: Array<{ __typename?: 'users', age?: string | null, id: any, image?: string | null, job?: string | null, name: string }> }> };